requirements.txt
attrs==21.4.0
factory-boy==3.2.1
Faker==13.0.0
greenlet==1.1.2
importlib-metadata==4.8.3
inflection==0.5.1
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.7
pytest==7.0.1
pytest-factoryboy==2.1.0
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==1.4.31
SQLAlchemy-Utils==0.38.2
tomli==1.2.3
typing_extensions==4.1.1
zipp==3.6.0
